requirements/base.txt
asgiref==3.6.0
astroid==2.15.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
cleo==2.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.3
crashtest==0.4.1
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.2
django-allauth==0.54.0
django-channels==0.7.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-dotenv==1.4.2
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.0.0
django-redis==5.2.0
django-shell-plus==1.1.7
django-silk==5.0.3
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
dulwich==0.21.3
environ==1.0
psycopg2==2.9.6
python-social-auth==0.3.6
social-auth-app-django==5.2.0
social-auth-core==4.4.2
whitenoise==6.4.0
argon2-cffi==21.3.0
channels-redis
channels[daphne]